==> 1_dataprep_1_columns.ipynb <==
# Data preparation - features examination

# imports
import pandas as pd
import pickle

# read files
df = pd.read_excel('.\meta\columns.xlsx')

df[df['categorical?']=="no"]

df['column']= df.apply(lambda x: str(x['code'])[2:-1]+ ": " + str(x['name'])[2:-1], axis=1)

df.drop("Unnamed: 5", axis=1, inplace=True)

df

df.to_pickle("./meta/columns")
